streaming_pipeline/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
kafka-python-ng
pandas
websockets==11.0.3
feedparser
